requirements.txt

# Core ML + Deep Learning
tensorflow==2.12.0
numpy==1.23.5
h5py==3.8.0

# Image handling
pillow==9.5.0

# Web app
streamlit==1.22.0
